move fliptrade into robot.js
